+++ doc/release-notes.rst
@@ -7,6 +7,7 @@ Upcoming Version
 
 
 * Rename `GEM_GGPT` to `GGPT` and add `GEM_GGPT` as an deprecated alias.
+* Rename `matched_data` to `powerplants` in `collection.py` and add `matched_data` as an deprecated alias.
 
 **Bug fixes**